Add a HttpPoller that remembers failures (#6)
At times it is reasonable to wait for a long (~minutes) time for resources
to become available - think large docker-composes on limited hardware.

This requires configuring the default poller with a fairly long timeout. Unfortunately
if you do that, and one of the resources is misconfigured and will never come up, you
will wait the entire long timeout for each test in your suite before aborting the entire
build. This can easily take you from a build taking ~10 minutes to over an hour.

The FailureCachingHttpPollingResource allows you to tolerate a long wait as long as the
resource becomes available eventually; but once you do time out, any subsequent tests will
fail eagerly instead of waiting the full timeout again.